Add read_buffer interface directly to pycbf (#3)
cbf_read_buffered_file is present in CBFlib, but undocumented in the HTML and
missing from the python bindings.

We added our own bindings to this in dxtbx to allow usage of CBFlib with
compressed CBF files (embedded in a .tar.gz or otherwise). This adds the
equivalent functionality directly into pycbf, as cbf_struct_handle.read_buffer.