Accept both bytes and str for pycbf functions (#4)
So that strings can be passed for filenames, field named, etc. Originally
committed in 647ffc, reapplied here in the non-nuweb branch.

This is designed to make it easier to transition away from
SWIG_PYTHON_STRICT_BYTE_CHAR. To that end, a convenience function
`pycbf.cbf2str` has been added, which will convert bytes to str, but
only if STRICT_BYTE_CHAR is turned on.

This also means that the regression tests now work on python 3, without
turning S_P_S_B_C on.